+++ wayland-scanner/src/client_gen.rs
@@ -96,6 +96,12 @@ fn generate_objects_for(interface: &Interface) -> TokenStream {
                 }
             }
 
+            impl std::hash::Hash for #iface_name {
+                fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
+                    self.id.hash(state)
+                }
+            }
+
             impl super::wayland_client::Proxy for #iface_name {
                 type Request = Request;
                 type Event = Event;