+++ wayland-scanner/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - Generated protocol object types now implement `Borrow<ObjectId>` and `Hash`
 
+#### Breaking changes
+
+- Bump bitflags to 2.0
+
 ## 0.30.0 -- 27/12/2022
 
 ## 0.30.0-beta.10